Extract barcode type to global scope
